moved logger to common
