+++ packages/api/src/infrastructure/weather/weather.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { WeatherProvider } from 'src/application/weather/interfaces/weather-provider.abstract';
 import { CacheAccessor } from 'src/common/interfaces/cache-service.interface';
 import { ProviderLogger } from 'src/common/interfaces/logger.interface';
+import { LoggerModule } from 'src/common/logger/logger.module';
 import { CacheModule } from 'src/infrastructure/cache/cache.module';
-import { LoggerModule } from 'src/infrastructure/logger/logger.module';
 import { OpenWeatherWeatherProvider } from 'src/infrastructure/weather/providers/openweather.provider';
 import { WeatherApiWeatherProvider } from 'src/infrastructure/weather/providers/weatherapi.provider';
 import { WeatherProviderCacheProxy } from 'src/infrastructure/weather/wrappers/weather-cache.proxy';
